Google user sign in

+++ Source/FaaS.MVC/wwwroot/src/actions/formsActions.js
@@ -20,9 +20,9 @@ export class FormsActions {
 		}
 	}
 
-	static fetchAll(projectId) {
+	static fetchAll(projectId, userId) {
 		return (dispatch) => {
-			apiClient.get(URL_ELEM + "?projectId="+ projectId)
+		    apiClient.get(URL_ELEM + "?projectId="+ projectId + "&userId=" + userId)
 				.then((res) => {
 					console.log(`[FETCH] ${COLL_NAME}: `, res.data);
 					dispatch ({type: COLL_TYPE.FetchSucc, payload: res.data});
@@ -34,9 +34,10 @@ export class FormsActions {
 		};
 	}
 
-	static del(id){
-		return (dispatch) => {
-			apiClient.delete(URL_ELEM + id)
+	static del(id, userId){
+	    return (dispatch) => {
+	        console.log("[ACTION] Form delete: ", id);
+		    apiClient.delete(URL_ELEM + `${id}?userId=${userId}`)
 				.then((res) => {
 					console.log(`[DELETE] ${COLL_NAME}: `, res);
 					dispatch({type: COLL_TYPE.DeleteSucc, payload: res.data});
@@ -49,9 +50,9 @@ export class FormsActions {
 		};
 	}
 
-	static update(element){
+	static update(form, userId){
 		return (dispatch) => {
-			apiClient.put(URL_ELEM, element)
+		    apiClient.put(URL_ELEM + `?userId=${userId}`, form)
 				.then(res => {
 					console.log(`[UPDATE] ${COLL_NAME}: `, res);
 					dispatch({
@@ -64,9 +65,9 @@ export class FormsActions {
 			});
 		}
 	}
-	static get(id){
+	static get(id, userId){
 		return (dispatch) => {
-			apiClient.get(URL_ELEM + id)
+			apiClient.get(URL_ELEM + id + "?userId=" + userId)
 				.then(res => {
 					console.log(`[GET] ${COLL_NAME}: `, res);
 					dispatch({
@@ -80,10 +81,10 @@ export class FormsActions {
 		}
 	}
 
-	static create(projectId, element)
+	static create(form)
 	{
 		return (dispatch) => {
-			apiClient.post(URL_ELEM, element)
+			apiClient.post(URL_ELEM, form)
 				.then(res => {
 					console.log(`[CREATE] ${COLL_NAME}: `, res);
 					dispatch({
