+++ Source/FaaS.MVC/wwwroot/src/actions/projectsActions.js
@@ -11,6 +11,7 @@ const COLL_TYPE = Projects;
 const FAIL_TYPE = COLL_TYPE.Fail;
 const COLL_NAME = "projects";
 const URL_ELEM = `/${COLL_NAME}/`;
+const URL_ELEM_QUERY = `/${COLL_NAME}?`;
 
 export class ProjectsActions {
 
@@ -22,7 +23,7 @@ export class ProjectsActions {
 
 	static fetchAll(userId) {
 		return (dispatch) => {
-			apiClient.get(URL_ELEM)
+			apiClient.get(URL_ELEM_QUERY + "userId=" + userId)
 				.then((res) => {
 					console.log(`[FETCH] ${COLL_NAME}: `, res);
 					dispatch({type: COLL_TYPE.FetchSucc, payload: res.data});
@@ -34,10 +35,10 @@ export class ProjectsActions {
 		};
 	}
 
-	static del(id){
+	static del(id, userId){
 		console.log("[ACTION] Project delete: ", id);
 		return (dispatch) => {
-			apiClient.delete(URL_ELEM + id)
+		    apiClient.delete(URL_ELEM + `${id}?userId=${userId}`)
 				.then((res) => {
 					console.log(`[DELETE] ${COLL_NAME}: `, res);
 					dispatch({type: COLL_TYPE.DeleteSucc, payload: res.data});
@@ -50,9 +51,9 @@ export class ProjectsActions {
 		};
 	}
 
-	static update(project){
+	static update(project, userId){
 		return (dispatch) => {
-			apiClient.put(URL_ELEM, project)
+		    apiClient.put(URL_ELEM + `?userId=${userId}`, project)
 				.then(res => {
 					console.log(`[UPDATE] ${COLL_NAME}: `, res);
 					dispatch({
@@ -65,9 +66,9 @@ export class ProjectsActions {
 			});
 		}
 	}
-	static get(id){
+	static get(id, userId){
 		return (dispatch) => {
-			apiClient.get(URL_ELEM + id)
+		    apiClient.get(URL_ELEM + `${id}?userId=${userId}`)
 				.then(res => {
 					console.log(`[GET] ${COLL_NAME}: `, res);
 					dispatch({
@@ -84,7 +85,7 @@ export class ProjectsActions {
 	static create(userId, project)
 	{
 		return (dispatch) => {
-			apiClient.post(URL_ELEM, project)
+		    apiClient.post(URL_ELEM + `?userId=${userId}`, project)
 				.then(res => {
 					console.log(`[CREATE] ${COLL_NAME}: `, res);
 					dispatch({
